Added internal stud to pipe side of socket
012345678901234567890123456789    456789    56789      6789012
The stud added is hollow, and not sized properly yet, this is
more of a proof of concept.

In the future, a look up table or another value for pipe wall
thickness should be passed, to size the stud properly for the
given pipe.  Lookup tables are feasable, since it's most likely
that Sched 40 or Sched 80 PVC pipe intend to be used.